@@ -13,14 +13,14 @@
         "archtechx/laravel-seo": "^0.10.1",
         "awcodes/filament-tiptap-editor": "^3.4",
         "blade-ui-kit/blade-icons": "^1.7",
-        "dedoc/scramble": "^0.11.25",
         "codeat3/blade-google-material-design-icons": "^1.19",
         "datlechin/filament-menu-builder": "^0.6.0",
+        "dedoc/scramble": "^0.11.25",
         "filament/filament": "^3.2",
         "filament/spatie-laravel-media-library-plugin": "^3.2",
         "haydenpierce/class-finder": "^0.5.3",
         "jeffgreco13/filament-breezy": "^2.4",
-        "laravel/framework": "^11.33",
+        "laravel/framework": "^11.34",
         "laravel/horizon": "^5.29",
         "laravel/sanctum": "^4.0",
         "laravel/scout": "^10.11",
@@ -40,7 +40,7 @@
         "barryvdh/laravel-debugbar": "^3.14",
         "barryvdh/laravel-ide-helper": "^3.2",
         "fakerphp/faker": "^1.24",
-        "friendsofphp/php-cs-fixer": "^3.64",
+        "friendsofphp/php-cs-fixer": "^3.65",
         "larastan/larastan": "^2.9",
         "laravel/pint": "^1.18",
         "laravel/sail": "^1.38",
