@@ -7,11 +7,9 @@ namespace App\Models;
 use App\Concerns\BelongsToElection;
 use App\Concerns\CanGroupByDataLevel;
 use App\Concerns\HasTemporaryTable;
-use App\Contracts\ClearsCache;
 use App\Contracts\TemporaryTable;
 use App\Enums\Area;
 use App\Enums\DataLevel;
-use App\Repositories\TurnoutRepository;
 use Database\Factories\TurnoutFactory;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
@@ -21,7 +19,7 @@ use Illuminate\Support\Collection;
 use Tpetry\QueryExpressions\Function\Aggregate\Sum;
 use Tpetry\QueryExpressions\Language\Alias;
 
-class Turnout extends Model implements TemporaryTable, ClearsCache
+class Turnout extends Model implements TemporaryTable
 {
     use BelongsToElection;
     use CanGroupByDataLevel;
@@ -189,10 +187,5 @@ class Turnout extends Model implements TemporaryTable, ClearsCache
         if ($this->hasAttribute('country_id') && $this->country_id) {
             return $this->loadMissing('country')->country->name;
         }
-    }
-
-    public function clearCache(int $electionId): bool
-    {
-        return TurnoutRepository::clearCache($electionId);
     }
 }
