@@ -10,11 +10,11 @@
         @endif
 
         <a
-            href="https://www.commitglobal.org"
+            href="https://www.code4.ro"
             target="_blank"
             rel="noopener"
             class="flex items-center w-full max-w-2xl gap-3 text-sm hover:text-blue-600 focus:text-blue-600 focus:outline-0 hover:underline focus:underline">
-            <x-icon-code4commit class="h-6 sm:h-8 shrink-0" />
+            <x-icon-code4romania class="h-6 sm:h-8 shrink-0" />
 
             <span>{{ __('app.banner') }}</span>
         </a>
